requirements.txt
# Helix Collective v14.5 - Frontend Requirements (Streamlit)
# This file is used by Streamlit Community Cloud
# Backend uses requirements-backend.txt (configured in railway.json)

# Core Framework
fastapi==0.115.0
python-dotenv==1.0.1

# API Clients (for backend communication)
httpx==0.28.0
aiohttp==3.11.0
requests==2.31.0

# Dashboard & Visualization
streamlit==1.40.0
plotly==5.18.0

# Utilities
pyyaml==6.0.1
pydantic==2.5.3

# Notion Integration (if needed for frontend)
notion-client==2.5.0



# MEGA Sync Dependency (Critical for Persistence)
mega.py


pycryptodome
